+++ src/components/Toolbar/Toolbar.test.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { fireEvent, render, screen } from '@testing-library/react';
-import { TestIds } from '../../../../constants';
+import { TestIds } from '../../constants';
 import { Toolbar } from './Toolbar';
 
 /**
