+++ src/components/index.ts
@@ -1,2 +1,3 @@
 export * from './CalendarPanel';
 export * from './MultiFieldEditor';
+export * from './Toolbar';
